+++ app/src/main/java/campus/tech/kakao/map/application/MyApplication.kt
@@ -1,6 +1,7 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.application
 
 import android.app.Application
+import campus.tech.kakao.map.R
 import com.kakao.vectormap.KakaoMapSdk
 import dagger.hilt.android.HiltAndroidApp
