+++ app/src/main/java/campus/tech/kakao/map/model/AppDatabase.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 import android.content.Context
 import androidx.room.Database
