+++ app/src/main/java/campus/tech/kakao/map/model/PlaceContract.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 object PlaceContract {
     const val TABLE_NAME = "places"
