+++ app/src/main/java/campus/tech/kakao/map/model/PlaceDao.kt
@@ -1,9 +1,10 @@
-package campus.tech.kakao.map
+package campus.tech.kakao.map.model
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
+import campus.tech.kakao.map.model.Place
 
 @Dao
 interface PlaceDao {
