+++ app/src/main/res/layout/activity_main.xml
@@ -2,18 +2,18 @@
 <layout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".MainActivity">
+    tools:context=".view.MainActivity">
 
     <data>
         <variable
             name="viewModel"
-            type="campus.tech.kakao.map.MainViewModel" />
+            type="campus.tech.kakao.map.viewmodel.MainViewModel" />
         <variable
             name="searchAdapter"
-            type="campus.tech.kakao.map.SearchAdapter" />
+            type="campus.tech.kakao.map.adapter.SearchAdapter" />
         <variable
             name="savedSearchAdapter"
-            type="campus.tech.kakao.map.SavedSearchAdapter" />
+            type="campus.tech.kakao.map.adapter.SavedSearchAdapter" />
 
         <import type="android.view.View"/>
     </data>
